% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAreplace.R
\name{NAreplace}
\alias{NAreplace}
\title{Replace NA with previous value}
\usage{
NAreplace(x)
}
\arguments{
\item{x}{Input vector}
}
\description{
This package helps to replace missing values in a vector with previous value
}
\examples{
NAreplace(c(NA, 1, 2, NA, NA, 3, NA, NA, 4, NA))

}
